renamed files and added todo comments
