+++ lib/models/color.js
