+++ lib/models/mood.js
