+++ lib/routes/auth.js
@@ -1,6 +1,6 @@
 const router = require('express').Router();
 const bodyParser = require('body-parser').json();
-const User = require('../models/user-schema');
+const User = require('../models/user');
 const token = require('../auth/token');
 const ensureAuth = require('../auth/ensure-auth')();
