+++ lib/routes/blocks.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const Block = require('../models/block-schema');
+const Block = require('../models/block');
 
 router
     .get('/block', (req, res, next) => {
