+++ lib/routes/colors.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const Color = require('../models/color-schema');
+const Color = require('../models/color');
 
 router
     .get('/color', (req, res, next) => {
