+++ lib/routes/index.js
@@ -2,7 +2,7 @@
 // this could become auth-route
 const express = require('express');
 const router = express.Router();
-const User = require('../models/user-schema');
+const User = require('../models/user');
 
 // GET /register
