+++ lib/routes/moods.js
@@ -2,8 +2,8 @@ const mongoose = require('mongoose');
 const ObjectId = mongoose.Types.ObjectId;
 const router = require('express').Router();
 const bodyParser = require('body-parser').json();
-const User = require('../models/user-schema');
-const Mood = require('../models/mood-schema');
+const User = require('../models/user');
+const Mood = require('../models/mood');
 
 router
     .get('/', (req, res, next) => {
