+++ lib/routes/users.js
@@ -1,6 +1,6 @@
 const mongoose = require('mongoose');
 const router = require('express').Router();
-const Mood = require('../models/mood-schema');
+const Mood = require('../models/mood');
 
 router
     .get('/all', (req, res, next) => {
