+++ test/weather-api.test.js
@@ -6,10 +6,10 @@ const assert = chai.assert;
 const app = require('../lib/app');
 const request = chai.request(app);
 const token = require('../lib/auth/token');
-const User = require('../lib/models/user-schema');
-const Mood = require('../lib/models/mood-schema');
-const Color = require('../lib/models/color-schema');
-const Block = require('../lib/models/block-schema');
+const User = require('../lib/models/user');
+const Mood = require('../lib/models/mood');
+const Color = require('../lib/models/color');
+const Block = require('../lib/models/block');
 const weatherApi = require('../lib/routes/weather-api');
 const apiKey = process.env.WEATHER_API_KEY;
 const requestProxy = require('request-promise');
